requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.27
zhipuai==2.1.5.20250106
jsonschema==4.23.0
requests==2.32.3
pytz==2025.1
urllib3==1.26.18
openai==1.12.0
httpx>=0.24.1
cryptography==44.0.0